[OSF noup] drivers: sensor: add stm32f401x temperature driver
There is an issue with this driver.

At present, there's no good way for board-level
configuration (specifically, the voltages on the V_REF- and V_REF+
external pins, which are needed to convert ADC values to volts) to be
made available portably to drivers.

Hack around this here by hard-coding reasonable defaults of 0V for
V_REF- and 3.3V for V_REF+.

Signed-off-by: Marti Bolivar <marti.bolivar@linaro.org>