[OSF toup] scripts: runner: dfu-util: support DT-based flashing for DfuSe
Enable DT support in the dfu-util flasher when the target is a
DfuSe (DFU + ST extensions) device. This requires untangling
DfuSe-specific options (currently, the default is 'leave', to
immediately start running after the flashing is done) from the actual
address itself. The address and options must be combined on the
dfu-util command line itself.

This lets us fix an unnecessary printline when flashing DfuSe devices
if 'leave' is given as an option. When that happens (and that's the
default behavior), there's no need to reset the device.

Signed-off-by: Marti Bolivar <marti@opensourcefoundries.com>