[OSF toup] scripts: runner: abstract openocd load/verify
The current runners don't support CONFIG_FLASH_LOAD_OFFSET when it is
present. This is unfortunate, since binaries built to be loaded by
bootloaders such as mcuboot need that to work properly. (Though note
that full mcuboot flashing support needs a bit of additional work,
since what's flashed at the load offset is a signed binary, not the
'raw' zephyr.bin.)

To prepare for adding this support for openocd, do some more
abstracting of how OPENOCD_LOAD_CMD and OPENOCD_VERIFY_CMD get
built. In addition to removing code from the individual board files
and moving it into the common include, this will make it easier to
globally change the address calculation to support
CONFIG_FLASH_LOAD_OFFSET later.

Signed-off-by: Marti Bolivar <marti@opensourcefoundries.com>