[OSF toup] net: lwm2m: dont release reply for duplicate block transfer
During normal use, a CoAP packet can be resent due to network congestion
and other causes.  During block transfer the LwM2M client checks to make
sure the block received is the one we expect and if not generates a
"duplicate" warning.  When this happened, we were releasing the reply
handler and when the correct block was received the client would generate
a "No handler" error.

To avoid releasing the reply handler too early, let's set the coap_reply
"user_data" field to 1.  Then, once the reply processing is complete we
can check the user_data field to be sure that it's ok to release the
reply handler.

Signed-off-by: Michael Scott <michael@opensourcefoundries.com>