[OSF fromlist] scripts: runner: nrfjprog: don't do full erase by default
By default, use a sector erase to flash boards with nrfjprog.py.

To allow getting the old behavior, add an --erase flag that works the
exact same way as the corresponding jlink.py argument: if present, a
full flash erase is done before programming. If absent, only a sector
erase is done.

Fixes #6147

Signed-off-by: Marti Bolivar <marti@opensourcefoundries.com>