[FIO toup] subsys: dfu: boot: mcuboot: introduce boot_request_erase()
This new public function can be used to erase the MCUboot metadata
vs. doing an entire erase of the flash via boot_erase_img_bank().

On some hardware with long erase times, performing an erase of
an entire bank takes too long under normal circumstances and a
function like this is needed.

Signed-off-by: Michael Scott <mike@foundries.io>