[FIO toup] subsys: dfu: add boot_invalidate_slot1()
This routine allows the caller to prevent subsequent boot from MCUboot
slot 1 without erasing the entire slot 1 area by zeroing out the
'magic' bytes MCUboot uses to check for the potential presence of a
valid image.

To do that, we use the native flash driver's page layout helpers
instead of the flash_map.h API, which doesn't provide what we need: we
require information about exactly the page in the slot 1 are which
contains the magic bytes, but flash_map.h only operates on entire
areas, which involves exactly the performance penalty we are trying to
avoid.

Note that using flash_area_get_sectors() is not an option here,
because we would need to allocate a (potentially large) array of
sectors to pass to it to find the required information, unlike
flash_get_page_info_by_offs(), which requires O(1) additional stack.

Signed-off-by: Marti Bolivar <marti@foundries.io>