[FIO toup] net: lwm2m: fix TLS handshake vs. poll() race
For UDP connections, socket API establishes TLS handshake on the
first send().  In the LwM2M engine we already start a poll()
on the socket at that point and it causes the handshake to abort.

Let's push off adding the socket to the poll() loop until after
the first successful send().

NOTE: This fixes DTLS connections for LwM2M library.

Signed-off-by: Michael Scott <mike@foundries.io>