[FIO toup] net: config: correct init wait for TIMEOUT setting
Currently, NET_CONFIG_INIT_TIMEOUT does very little to alter
the actual timeout for network init.  If you leave it at the
default of 90, network init will timeout around a minute and
30 seconds.
If you set it to 30, network init will timeout around a
minute and 30 seconds.  This has to do with how the loop delay
and the total loop count are calculated.

Let's fix the calculations so that we get a more accurate
timeout setting.

Signed-off-by: Michael Scott <mike@foundries.io>