[FIO toup] net: lwm2m: retransmit work should be part of messages
The logic for lwm2m_ctx containing the delay_work structure used to
process outstanding LwM2M messages is flawed.  If at any point we
cancel the work while there are outstanding messages to be processed
they will be orphaned.

Instead, attach the delay_work to the message itself.  Each time we
send a message, start the pending timer.  If the timer fires off we
know exactly which message needs to be resent and we kick off the
delay_work again.  This way we elminate the flawed logic entirely.

Also remove the now useless function: lwm2m_engine_context_init()

Signed-off-by: Michael Scott <mike@foundries.io>