[FIO toup] Revert "misc/dlist: Swap insertion API for a faster one"
This commit seems to be causing a regression in the kernel/timeout
code where the device will endlessly loop in the timeout_list
looking for a place to insert the new timeout.  This apparently
is caused by a corrupted timeout_list where a portion is looped.

Testing showed that we could successfully OTA and reboot 136
times with this reverted.  Where prior, any number of devices
would get stuck in this loop during the initial portion of
initialization after entry into the main() function.

More testing and analysis is needed.

This reverts commit eda4c027dadf8a8e62da1342e855aff91285a3ab.

Signed-off-by: Michael Scott <mike@foundries.io>