[FIO fromlist] boards: arm: nrf52840-based: Free up flash room for sample apps
The current flash configurations for all nRF52840's in Zephyr is
VERY constrained when it comes to allowing samples any space for
storage or custom areas.  It only leaves the last 4 pages of flash
for "storage".

The nRF52840 is also capable of using OpenThread which defaults
to using the last 4 pages of flash for storing OpenThread-related
network data.

This means that while using OpenThread under any configuration
designed to use mcuboot partition slots, there is no space left
over for storage of any kind.

Let's reduce the scratch partition size to a reasonable 72k
(down from 120k) and make room for storage, OpenThread and
a bit extra which can be used later when more use cases arise.

Signed-off-by: Michael Scott <mike@foundries.io>