[FIO fromtree] boards: arm: particle_*: remove SPI default enablement
The nRF52840-based Particle.io boards have an external flash
module that is connected via SPI interface.  Originally,
the goal was to place the flash part into deep power down
mode and save 9 uA in sleep mode, but this was never
implemented in a board initialization file.

Let's remove the default y behavior for SPI as each sample
will turn it on if needed, and it can cause certain out of
tree samples to behave badly (where CONFIG_MULTITHREADING=n).

Long term: we need a better solution for handling samples
that don't fit the requirements for certain drivers, but
for now this will clean up some board usability issues.

Signed-off-by: Michael Scott <mike@foundries.io>